+++ src/autoscaling/autoscaled_pool.js
@@ -83,12 +83,12 @@ const DEFAULT_OPTIONS = {
  *   To abort a run, use the [`abort()`](#AutoscaledPool+abort) method.
  *
  * @param {Number} [options.minConcurrency=1]
- *   Minimum number of tasks running in parallel.
+ *   The minimum number of tasks running in parallel.
  *
  *   *WARNING:* If you set this value too high with respect to the available system memory and CPU, your code might run extremely slow or crash.
  *   If you're not sure, just keep the default value and the concurrency will scale up automatically.
  * @param {Number} [options.maxConcurrency=1000]
- *   Maximum number of tasks running in parallel.
+ *   The maximum number of tasks running in parallel.
  * @param {Number} [options.desiredConcurrencyRatio=0.95]
  *   Minimum level of desired concurrency to reach before more scaling up is allowed.
  * @param {Number} [options.scaleUpStepRatio=0.05]
@@ -148,8 +148,6 @@ class AutoscaledPool {
         checkParamOrThrow(snapshotterOptions, 'options.snapshotterOptions', 'Maybe Object');
 
         // Configurable properties.
-        this.maxConcurrency = maxConcurrency;
-        this.minConcurrency = minConcurrency;
         this.desiredConcurrencyRatio = desiredConcurrencyRatio;
         this.scaleUpStepRatio = scaleUpStepRatio;
         this.scaleDownStepRatio = scaleDownStepRatio;
@@ -161,9 +159,11 @@ class AutoscaledPool {
         this.isTaskReadyFunction = isTaskReadyFunction;
 
         // Internal properties.
+        this._maxConcurrency = maxConcurrency;
+        this._minConcurrency = minConcurrency;
+        this._desiredConcurrency = this._minConcurrency;
+        this._currentConcurrency = 0;
         this.isStopped = false;
-        this.desiredConcurrency = this.minConcurrency;
-        this.currentConcurrency = 0;
         this.lastLoggingTime = 0;
         this.resolve = null;
         this.reject = null;
@@ -177,24 +177,79 @@ class AutoscaledPool {
         this.systemStatus = new SystemStatus(ssoCopy);
     }
 
+
     /**
-     * Overrides max concurrency configuration.
-     *
-     * @param {Number} maxConcurrency
+     * @private
      */
     setMaxConcurrency(maxConcurrency) {
-        this.maxConcurrency = maxConcurrency;
+        log.deprecated('AutoscaledPool.setMaxConcurrency() is deprecated, use the "maxConcurrency" property instead');
+        this._maxConcurrency = maxConcurrency;
     }
 
     /**
-     * Overrides min concurrency configuration.
+     * @private
+     */
+    setMinConcurrency(minConcurrency) {
+        log.deprecated('AutoscaledPool.setMaxConcurrency() is deprecated, use the "maxConcurrency" property instead');
+        this._minConcurrency = minConcurrency;
+    }
+
+
+    /**
+     * Gets the minimum number of tasks running in parallel.
+     */
+    get minConcurrency() {
+        return this._minConcurrency;
+    }
+
+    /**
+     * Sets the minimum number of tasks running in parallel.
      *
      * *WARNING:* If you set this value too high with respect to the available system memory and CPU, your code might run extremely slow or crash.
      * If you're not sure, just keep the default value and the concurrency will scale up automatically.
-     * @param {Number} minConcurrency
      */
-    setMinConcurrency(minConcurrency) {
-        this.minConcurrency = minConcurrency;
+    set minConcurrency(value) {
+        this._minConcurrency = value;
+    }
+
+    /**
+     * Gets the maximum number of tasks running in parallel.
+     */
+    get maxConcurrency() {
+        return this._maxConcurrency;
+    }
+
+    /**
+     * Sets the maximum number of tasks running in parallel.
+     */
+    set maxConcurrency(value) {
+        this._maxConcurrency = value;
+    }
+
+    /**
+     * Gets the desired concurrency for the pool,
+     * which is an estimated number of parallel tasks that the system can currently support.
+     */
+    get desiredConcurrency() {
+        return this._desiredConcurrency;
+    }
+
+    /**
+     * Sets the desired concurrency for the pool, i.e. the number of tasks that should be running
+     * in parallel if there's large enough supply of tasks.
+     */
+    set desiredConcurrency(value) {
+        if (!(this._minConcurrency <= value && value <= this._maxConcurrency)) {
+            throw new Error(`The desired concurrency must be an integer from ${this._minConcurrency} to ${this._maxConcurrency}`);
+        }
+        this._desiredConcurrency = value;
+    }
+
+    /**
+     * Gets the the number of parallel tasks currently running in the pool.
+     */
+    get currentConcurrency() {
+        return this._currentConcurrency;
     }
 
     /**
@@ -276,7 +331,7 @@ class AutoscaledPool {
             }
 
             const interval = setInterval(() => {
-                if (this.currentConcurrency <= 0) {
+                if (this._currentConcurrency <= 0) {
                     // Clean up timeout and interval to prevent process hanging.
                     if (timeout) clearTimeout(timeout);
                     clearInterval(interval);
@@ -316,12 +371,12 @@ class AutoscaledPool {
         // - we are already querying for a task.
         if (this.queryingIsTaskReady) return done();
         // - we would exceed desired concurrency.
-        if (this.currentConcurrency >= this.desiredConcurrency) return done();
+        if (this._currentConcurrency >= this._desiredConcurrency) return done();
         // - system is overloaded now and we are at or above minConcurrency
         const currentStatus = this.systemStatus.getCurrentStatus();
         const { isSystemIdle } = currentStatus;
-        if (!isSystemIdle && this.currentConcurrency >= this.minConcurrency) {
-            log.debug('AutoscaledPool: Task will not run, system is overloaded.', currentStatus);
+        if (!isSystemIdle && this._currentConcurrency >= this._minConcurrency) {
+            log.perf('AutoscaledPool: Task will not be run. System is overloaded.', currentStatus);
             return done();
         }
         // - a task is ready.
@@ -347,7 +402,7 @@ class AutoscaledPool {
 
         try {
             // Everything's fine. Run task.
-            this.currentConcurrency++;
+            this._currentConcurrency++;
             // Try to run next task to build up concurrency,
             // but defer it so it doesn't create a cycle.
             setImmediate(this._maybeRunTask);
@@ -357,7 +412,7 @@ class AutoscaledPool {
 
             // Execute the current task.
             await this.runTaskFunction();
-            this.currentConcurrency--;
+            this._currentConcurrency--;
             // Run task after the previous one finished.
             setImmediate(this._maybeRunTask);
         } catch (err) {
@@ -386,10 +441,10 @@ class AutoscaledPool {
         const systemStatus = this.systemStatus.getHistoricalStatus();
         const { isSystemIdle } = systemStatus;
         // - we're not already at max concurrency.
-        const weAreNotAtMax = this.desiredConcurrency < this.maxConcurrency;
+        const weAreNotAtMax = this._desiredConcurrency < this._maxConcurrency;
         // - current concurrency reaches at least the given ratio of desired concurrency.
-        const minCurrentConcurrency = Math.floor(this.desiredConcurrency * this.desiredConcurrencyRatio);
-        const weAreReachingDesiredConcurrency = this.currentConcurrency >= minCurrentConcurrency;
+        const minCurrentConcurrency = Math.floor(this._desiredConcurrency * this.desiredConcurrencyRatio);
+        const weAreReachingDesiredConcurrency = this._currentConcurrency >= minCurrentConcurrency;
 
         if (isSystemIdle && weAreNotAtMax && weAreReachingDesiredConcurrency) this._scaleUp(systemStatus);
 
@@ -397,7 +452,7 @@ class AutoscaledPool {
         // - the system has been overloaded lately.
         const isSystemOverloaded = !isSystemIdle;
         // - we're over min concurrency.
-        const weAreNotAtMin = this.desiredConcurrency > this.minConcurrency;
+        const weAreNotAtMin = this._desiredConcurrency > this._minConcurrency;
 
         if (isSystemOverloaded && weAreNotAtMin) this._scaleDown(systemStatus);
 
@@ -407,8 +462,8 @@ class AutoscaledPool {
             if (now > this.lastLoggingTime + this.loggingIntervalMillis) {
                 this.lastLoggingTime = now;
                 log.info('AutoscaledPool state', {
-                    currentConcurrency: this.currentConcurrency,
-                    desiredConcurrency: this.desiredConcurrency,
+                    currentConcurrency: this._currentConcurrency,
+                    desiredConcurrency: this._desiredConcurrency,
                     systemStatus,
                 });
             }
@@ -426,11 +481,11 @@ class AutoscaledPool {
      * @ignore
      */
     _scaleUp(systemStatus) {
-        const step = Math.ceil(this.desiredConcurrency * this.scaleUpStepRatio);
-        this.desiredConcurrency = Math.min(this.maxConcurrency, this.desiredConcurrency + step);
+        const step = Math.ceil(this._desiredConcurrency * this.scaleUpStepRatio);
+        this._desiredConcurrency = Math.min(this._maxConcurrency, this._desiredConcurrency + step);
         log.debug('AutoscaledPool: scaling up', {
-            oldConcurrency: this.desiredConcurrency - step,
-            newConcurrency: this.desiredConcurrency,
+            oldConcurrency: this._desiredConcurrency - step,
+            newConcurrency: this._desiredConcurrency,
             systemStatus,
         });
     }
@@ -443,11 +498,11 @@ class AutoscaledPool {
      * @ignore
      */
     _scaleDown(systemStatus) {
-        const step = Math.ceil(this.desiredConcurrency * this.scaleUpStepRatio);
-        this.desiredConcurrency = Math.max(this.minConcurrency, this.desiredConcurrency - step);
+        const step = Math.ceil(this._desiredConcurrency * this.scaleUpStepRatio);
+        this._desiredConcurrency = Math.max(this._minConcurrency, this._desiredConcurrency - step);
         log.debug('AutoscaledPool: scaling down', {
-            oldConcurrency: this.desiredConcurrency + step,
-            newConcurrency: this.desiredConcurrency,
+            oldConcurrency: this._desiredConcurrency + step,
+            newConcurrency: this._desiredConcurrency,
             systemStatus,
         });
     }
@@ -462,7 +517,7 @@ class AutoscaledPool {
      */
     async _maybeFinish() {
         if (this.queryingIsFinished) return;
-        if (this.currentConcurrency > 0) return;
+        if (this._currentConcurrency > 0) return;
 
         this.queryingIsFinished = true;
         try {